move colors.ts to private folder
