+++ packages/cli-kit/src/public/node/output.ts
@@ -1,7 +1,7 @@
 /* eslint-disable no-console */
 import {isUnitTest, isVerbose} from './context/local.js'
 import {PackageManager} from './node-package-manager.js'
-import colors from './colors.js'
+import colors from '../../private/node/colors.js'
 import {
   ColorContentToken,
   CommandContentToken,
